New Phase 1 feature: "Same thing as before but ABOVE-median rent, rather than below-median."
To minimize redundant code, I built a toggle into compare_rent_and_growth that defaults to checking for below-median rents but adds an "above-median" keyword. I am concerned that the function is getting too long, but the percentile ranking of ZIP codes draws on information generated by earlier calculations, so it seems most efficient to do all the work in this function. I'm open to recommendations on how or if to refactor this function!