New Phase 1 feature: "Ranking candidate zip codes in terms of price-to-rent ratio / perhaps also secondarily ranking by something else (e.g., rent growth, home value, etc.)?"
Implemented compare_price_to_rent(). This required updating zillow_data_parser() to function with both rent and home-value data, so I added a "metric" parameter to specify which is being parsed. (Sidenote: Zilow has data tailored to how many bedrooms a house has, which we could try using.) I'm using the same try-except code to check if my aggregation function has been called, so I'm thinking it would make more sense to ensure this column has been generated in advance ... But I'm not sure what the best structure for doing that is, so I'd like to talk it through.